+++ tests/testthat/test_fitting.R
@@ -20,12 +20,12 @@ test_that("Test ABfit with default options", {
   mrs_data <- readRDS("abfit_sim_mrs_data.rds")
   sim_res  <- readRDS("abfit_sim_res.rds")
   
-  fit_res  <- fit_mrs(mrs_data, sim_res$basis, method = "abfit", time = FALSE)
+  fit_res  <- fit_mrs(mrs_data, sim_res$basis, method = "abfit", time = FALSE,
+                      progress = "none")
   
   expect_equal_to_reference(fit_res, "abfit_res_default.rds", tolerance = 1e-1)
   ref_val <- 0.4465198
-  expect_equal(fit_res$res_tab$res.deviance, ref_val, tolerance = 1e-3)
-  
+  expect_equal(fit_res$res_tab$res.deviance, ref_val, tolerance = 1.5e-3)
 })
 
 test_that("Test ABfit without iterative optimisation", {
@@ -37,7 +37,7 @@ test_that("Test ABfit without iterative optimisation", {
   
   # undo the large phase error to avoid FWHM measurement warnings
   fit_res  <- fit_mrs(mrs_data %>% phase(-130), sim_res$basis, method = "abfit",
-                      opts = opts, time = FALSE)
+                      opts = opts, time = FALSE, progress = "none")
   
   expect_equal_to_reference(fit_res, "abfit_res_no_optim.rds", tolerance = 1e-5)
 })
@@ -49,7 +49,7 @@ test_that("Test ABfit coarse-fitting steps only", {
   # don't run the fine fitting step
   opts <- abfit_opts(maxiters = 0)
   fit_res  <- fit_mrs(mrs_data, sim_res$basis, method = "abfit", opts = opts,
-                      time = FALSE)
+                      time = FALSE, progress = "none")
   
   # this test is a bit more sensitive to differences between platforms
   expect_equal_to_reference(fit_res, "abfit_res_coarse.rds", tolerance = 1e-1)
@@ -66,7 +66,7 @@ test_that("Test ABfit fine-fitting only", {
   
   # undo the large phase error to avoid FWHM measurement warnings
   fit_res  <- fit_mrs(mrs_data %>% phase(-120), sim_res$basis, method = "abfit",
-                      opts = opts, time = FALSE)
+                      opts = opts, time = FALSE, progress = "none")
   
   expect_equal_to_reference(fit_res, "abfit_res_fine.rds", tolerance = 1e-6)
 })